pixel-prompt-google/requirements.txt
Pillow==9.5.0
annotated_types==0.7.0
anyio==4.9.0
boto3==1.37.30
botocore==1.37.30
cachetools==5.5.2
certifi==2025.1.31
charset_normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
google-auth==2.38.0
google-genai==1.10.0
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
lxml==5.3.2
openai==1.72.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
python-dateutil==2.9.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
websockets==15.0.1